+++ src/main/scala/edu/ie3/simona/service/Data.scala
@@ -18,6 +18,8 @@ import squants.{Temperature, Velocity}
 import squants.energy.{Kilowatts, Power}
 import tech.units.indriya.ComparableQuantity
 
+import java.time.ZonedDateTime
+import scala.collection.SortedMap
 import scala.jdk.OptionConverters.RichOptional
 import scala.reflect.ClassTag
 import scala.util.{Failure, Success, Try}
@@ -379,6 +381,16 @@ object Data {
         dirIrr: Irradiance,
         temp: Temperature,
         windVel: Velocity,
+    ) extends SecondaryData
+
+    /** Container class for weather data series over some time interval for
+      * certain coordinate.
+      *
+      * @param series
+      *   The weather time series consisting of [[WeatherData]].
+      */
+    final case class WeatherSeriesData(
+        series: SortedMap[ZonedDateTime, WeatherData]
     ) extends SecondaryData
 
     final case class WholesalePrice(price: ComparableQuantity[EnergyPrice])