+++ src/main/scala/edu/ie3/simona/service/weather/WeatherService.scala
@@ -15,24 +15,27 @@ import edu.ie3.simona.agent.participant.ParticipantAgent.{
 import edu.ie3.simona.config.InputConfig
 import edu.ie3.simona.exceptions.InitializationException
 import edu.ie3.simona.exceptions.WeatherServiceException.InvalidRegistrationRequestException
+import edu.ie3.simona.ontology.messages.ServiceMessage.{
+  SecondaryServiceRegistrationMessage,
+  ServiceRegistrationMessage,
+}
+import edu.ie3.simona.service.Data.SecondaryData.WeatherSeriesData
 import edu.ie3.simona.service.ServiceStateData.{
   InitializeServiceStateData,
   ServiceBaseStateData,
 }
 import edu.ie3.simona.service.SimonaService
-import edu.ie3.simona.ontology.messages.ServiceMessage.{
-  SecondaryServiceRegistrationMessage,
-  ServiceRegistrationMessage,
-}
 import edu.ie3.simona.service.weather.WeatherSource.WeightedCoordinates
-import edu.ie3.simona.util.{Coordinate, SimonaConstants}
 import edu.ie3.simona.util.TickUtil.RichZonedDateTime
+import edu.ie3.simona.util.{Coordinate, SimonaConstants}
 import edu.ie3.util.scala.collection.immutable.SortedDistinctSeq
 import org.apache.pekko.actor.typed.ActorRef
 import org.apache.pekko.actor.typed.scaladsl.ActorContext
 import org.slf4j.Logger
+import squants.Time
 
 import java.time.ZonedDateTime
+import scala.collection.immutable.SortedMap
 import scala.util.{Failure, Success, Try}
 
 /** Weather Service is responsible to register other actors that require weather
@@ -43,28 +46,51 @@ import scala.util.{Failure, Success, Try}
   */
 object WeatherService extends SimonaService {
 
-  override type S = WeatherInitializedStateData
+  override type S = WeatherBaseStateData
 
-  /** @param weatherSource
-    *   weather source to receive information from
-    * @param coordsToActorRefMap
-    *   mapping of the requested coords to their receiving actor references
-    * @param weightedWeatherCoordinates
-    *   Maps the requested agent coordinates onto the surrounding weather
-    *   coordinates, including their weighting
-    * @param maybeNextActivationTick
-    *   the next tick, when this actor is triggered by scheduler
-    * @param activationTicks
-    *   sorted set of ticks, that yet have been sent to the scheduler (w\o next
-    *   tick)
+  /** Data required for registration with the weather service.
+    *
+    * @param coordinate
+    *   The coordinate to register weather data for.
+    * @param dataType
+    *   The type of weather data to register for.
     */
-  final case class WeatherInitializedStateData(
-      weatherSource: WeatherSource,
-      coordsToActorRefMap: Map[Coordinate, Vector[
+  final case class WeatherRegistrationData(
+      coordinate: Coordinate,
+      dataType: WeatherDataType,
+  )
+
+  /** Container storing registered actors for a coordinate.
+    *
+    * @param registeredActors
+    *   A map of weather data type to registered actors.
+    * @param coordinateWeights
+    *   Weights mapping surrounding coordinates onto the registered coordinate.
+    */
+  final case class CoordinateData(
+      registeredActors: Map[WeatherDataType, Set[
         ActorRef[ParticipantAgent.Request]
-      ]] = Map.empty,
-      weightedWeatherCoordinates: Map[Coordinate, WeightedCoordinates] =
-        Map.empty,
+      ]],
+      coordinateWeights: WeightedCoordinates,
+  )
+
+  /** State data of an initialized weather service.
+    *
+    * @param weatherSource
+    *   The weather source to retrieve information from.
+    * @param coordinateData
+    *   A map of the requested coords to their receiving actor references.
+    * @param maybeNextActivationTick
+    *   The next tick at which the service wants to be activated.
+    * @param activationTicks
+    *   A sorted set of ticks, that yet have been sent to the scheduler (w\o
+    *   next tick).
+    * @param amountOfInterpolationCoords
+    *   The amount of coordinates to be interpolated for any agent coordinate.
+    */
+  final case class WeatherBaseStateData(
+      weatherSource: WeatherSource,
+      coordinateData: Map[Coordinate, CoordinateData] = Map.empty,
       maybeNextActivationTick: Option[Long],
       activationTicks: SortedDistinctSeq[Long] = SortedDistinctSeq.empty,
       amountOfInterpolationCoords: Int = 4,
@@ -83,7 +109,7 @@ object WeatherService extends SimonaService {
 
   override def init(
       initServiceData: InitializeServiceStateData
-  )(using log: Logger): Try[(WeatherInitializedStateData, Option[Long])] =
+  )(using log: Logger): Try[(WeatherBaseStateData, Option[Long])] =
     initServiceData match {
       case InitWeatherServiceStateData(
             sourceDefinition,
@@ -104,7 +130,7 @@ object WeatherService extends SimonaService {
             .toSeq
         ).pop
 
-        val weatherInitializedStateData = WeatherInitializedStateData(
+        val weatherInitializedStateData = WeatherBaseStateData(
           weatherSource,
           activationTicks = furtherActivationTicks,
           maybeNextActivationTick = maybeNextTick,
@@ -123,31 +149,22 @@ object WeatherService extends SimonaService {
         )
     }
 
-  /** Handle a request to register for information from this service
-    *
-    * @param registrationMessage
-    *   registration message to handle
-    * @param serviceStateData
-    *   current state data of the actor
-    * @return
-    *   the service stata data that should be used in the next state (normally
-    *   with updated values)
-    */
   override def handleRegistrationRequest(
       registrationMessage: ServiceRegistrationMessage
   )(using
-      serviceStateData: WeatherInitializedStateData,
+      serviceStateData: WeatherBaseStateData,
       ctx: ActorContext[Message],
-  ): Try[WeatherInitializedStateData] =
+  ): Try[WeatherBaseStateData] =
     registrationMessage match {
       case SecondaryServiceRegistrationMessage(
             agentToBeRegistered,
-            coordinate: Coordinate,
+            WeatherRegistrationData(coordinate, dataType),
           ) =>
         Success(
           handleRegistrationRequest(
             agentToBeRegistered,
             coordinate,
+            dataType,
           )
         )
       case invalidMessage =>
@@ -159,13 +176,15 @@ object WeatherService extends SimonaService {
         )
     }
 
-  /** Try to register the sending agent with its latitude and longitude values
-    * for weather provision.
+  /** Try to register the sending agent with its coordinate and weather data
+    * type for weather provision.
     *
     * @param agentToBeRegistered
     *   The agent that wants to be registered.
     * @param coordinate
     *   The coordinate of the agent to be registered.
+    * @param dataType
+    *   The weather data type that the agent wants to receive.
     * @param serviceStateData
     *   The current service state data of this service.
     * @return
@@ -175,10 +194,11 @@ object WeatherService extends SimonaService {
   private def handleRegistrationRequest(
       agentToBeRegistered: ActorRef[ParticipantAgent.Request],
       coordinate: Coordinate,
+      dataType: WeatherDataType,
   )(using
-      serviceStateData: WeatherInitializedStateData,
+      serviceStateData: WeatherBaseStateData,
       ctx: ActorContext[Message],
-  ): WeatherInitializedStateData = {
+  ): WeatherBaseStateData = {
     ctx.log.debug(
       "Received weather registration from {} for [Lat:{}, Long:{}]",
       agentToBeRegistered.path.name,
@@ -191,7 +211,7 @@ object WeatherService extends SimonaService {
       .map(RegistrationSuccessfulMessage(ctx.self, _))
       .getOrElse(RegistrationFailedMessage(ctx.self))
 
-    serviceStateData.coordsToActorRefMap.get(coordinate) match {
+    serviceStateData.coordinateData.get(coordinate) match {
       case None =>
         /* The coordinate itself is not known yet. Try to figure out, which weather coordinates are relevant */
         serviceStateData.weatherSource.getWeightedCoordinates(
@@ -201,15 +221,16 @@ object WeatherService extends SimonaService {
           case Success(weightedCoordinates) =>
             agentToBeRegistered ! registrationResponse
 
+            val coordinateData = CoordinateData(
+              registeredActors = Map(dataType -> Set(agentToBeRegistered)),
+              coordinateWeights = weightedCoordinates,
+            )
+
             /* Enhance the mapping from agent coordinate to requesting actor's ActorRef as well as the necessary
              * weather coordinates for later averaging. */
             serviceStateData.copy(
-              coordsToActorRefMap =
-                serviceStateData.coordsToActorRefMap + (coordinate -> Vector(
-                  agentToBeRegistered
-                )),
-              weightedWeatherCoordinates =
-                serviceStateData.weightedWeatherCoordinates + (coordinate -> weightedCoordinates),
+              coordinateData =
+                serviceStateData.coordinateData + (coordinate -> coordinateData)
             )
           case Failure(exception) =>
             ctx.log.error(
@@ -220,51 +241,39 @@ object WeatherService extends SimonaService {
             serviceStateData
         }
 
-      case Some(actorRefs) if !actorRefs.contains(agentToBeRegistered) =>
-        // coordinate is already known (= we have data for it), but this actor is not registered yet
-        agentToBeRegistered ! registrationResponse
+      case Some(coordinateData) =>
+        val registeredActors =
+          coordinateData.registeredActors.getOrElse(dataType, Set.empty)
+
+        if registeredActors.contains(agentToBeRegistered) then
+          ctx.log.warn(
+            "Sending actor {} is already registered",
+            agentToBeRegistered,
+          )
+        else agentToBeRegistered ! registrationResponse
+
+        val adaptedCoordinateData = coordinateData.copy(registeredActors =
+          coordinateData.registeredActors +
+            (dataType -> registeredActors.incl(agentToBeRegistered))
+        )
 
         serviceStateData.copy(
-          coordsToActorRefMap =
-            serviceStateData.coordsToActorRefMap + (coordinate -> (actorRefs :+ agentToBeRegistered))
+          coordinateData =
+            serviceStateData.coordinateData + (coordinate -> adaptedCoordinateData)
         )
 
-      case Some(actorRefs) if actorRefs.contains(agentToBeRegistered) =>
-        // actor is already registered, do nothing
-        ctx.log.warn(
-          "Sending actor {} is already registered",
-          agentToBeRegistered,
-        )
-        serviceStateData
-
-      case _ =>
-        // actor is not registered, and we don't have data for it
-        // inform the agentToBeRegistered that the registration failed as we don't have data for it
-        agentToBeRegistered ! RegistrationFailedMessage(ctx.self)
-        serviceStateData
     }
   }
 
-  /** Send out the information to all registered recipients
-    *
-    * @param tick
-    *   current tick data should be announced for
-    * @param serviceStateData
-    *   the current state data of this service
-    * @return
-    *   the service stata data that should be used in the next state (normally
-    *   with updated values) together with the completion message that is sent
-    *   in response to the trigger that was sent to start this announcement
-    */
   override protected def announceInformation(tick: Long)(using
-      serviceStateData: WeatherInitializedStateData,
+      serviceStateData: WeatherBaseStateData,
       ctx: ActorContext[Message],
-  ): (WeatherInitializedStateData, Option[Long]) = {
+  ): (WeatherBaseStateData, Option[Long]) = {
 
     /* Pop the next activation tick and update the state data */
     val (
       maybeNextTick: Option[Long],
-      updatedStateData: WeatherInitializedStateData,
+      updatedStateData: WeatherBaseStateData,
     ) = {
       val (nextTick, remainderTicks) = serviceStateData.activationTicks.pop
       (nextTick, serviceStateData.copy(activationTicks = remainderTicks))
@@ -273,26 +282,71 @@ object WeatherService extends SimonaService {
     // get the weather and send it to the subscribed agents
     // no sanity check needed here as we can assume that we always have weather available
     // when we announce it. Otherwise, the registration would have failed already!
-    updatedStateData.weatherSource
-      .getWeather(tick, updatedStateData.weightedWeatherCoordinates)
-      .foreach { case coordinate -> weatherResult =>
-        updatedStateData.coordsToActorRefMap
-          .get(coordinate)
-          .foreach(recipients =>
-            recipients.foreach(
-              _ ! DataProvision(
-                tick,
-                ctx.self,
-                weatherResult,
-                maybeNextTick,
-              )
+    updatedStateData.coordinateData.foreach { case (_, coordinateData) =>
+      val coordinateWeights = coordinateData.coordinateWeights
+
+      coordinateData.registeredActors.foreach { case (dataType, actors) =>
+        val weatherData = dataType match {
+          case WeatherDataType.Current =>
+            updatedStateData.weatherSource.getWeather(tick, coordinateWeights)
+          case WeatherDataType.CurrentAndForecast(length, interval) =>
+            val endTick = tick + length.toSeconds.toLong
+            // weather time series is forwarded without adding error
+            val series = updatedStateData.weatherSource.getWeather(
+              tick,
+              endTick,
+              coordinateWeights,
             )
+            WeatherSeriesData(reduceTimeSeriesResolution(series, interval))
+        }
+
+        actors.foreach {
+          _ ! DataProvision(
+            tick,
+            ctx.self,
+            weatherData,
+            maybeNextTick,
           )
+        }
       }
+    }
 
     (
       updatedStateData,
       maybeNextTick,
     )
   }
+
+  /** Reduces the resolution of given time series to at least given resolution
+    * by removing elements from the time series.
+    *
+    * @param timeSeries
+    *   The time series to adapt.
+    * @param resolution
+    *   The time resolution to aim for.
+    * @tparam T
+    *   The type of time series data.
+    * @return
+    *   The adapted time series.
+    */
+  def reduceTimeSeriesResolution[T](
+      timeSeries: SortedMap[ZonedDateTime, T],
+      resolution: Time,
+  ): SortedMap[ZonedDateTime, T] = {
+    val resolutionSeconds = resolution.toSeconds.toLong
+
+    timeSeries.foldLeft(timeSeries) { case (result, (it, _)) =>
+      result.maxBefore(it) match {
+        case Some((last, _)) =>
+          if last.plusSeconds(resolutionSeconds).isAfter(it) then
+            // interval from last to current key is too short
+            result.removed(it)
+          else result
+        case None =>
+          // no data before the current key, keep it
+          result
+      }
+    }
+  }
+
 }